[semver:patch] Fail the continuation request on non-200
This means the job will actually fail, which is much clearer.

This is a but cumbersome, because curl doesn't actually fail on
negative responses. Even worse, with --fail it does, but doesn't show
the response body, which isn't particularly helpful.

The way this works is we redirect the response body to stderr. Then we
write the HTTP response code to stdout, and compare it against 200.
Bash will return a status code of 1 if it doesn't match, but still
print stderr in any case.