Pass the config file through files rather than the env
This avoids bumping into the ARG_MAX limit of a few hundred kB which
can be inside the environment at a given time. We officially support
configs up to 3MB (though that's problematic in its own right), but
his gets us at least much closer to that.

The big problem we were facing was that jq assumes all its input will
be more or less valid JSON, but we actually just want to read in a big
string from a file. This string also needs to be escaped to actually
be a valid JSON string.

The solution is to adapt a two-pass approach, first invoking jq just
to escape the config file as a JSON string and store it in another
temporary file. Then read it in in the main jq invocation with
slurpfile, which reads in the lines into an array, which we can then
join again with newlines.

It's a bit roundabout, but works.